Changes to make data uint32

+++ baseband/src/chip_helpers/chip_helpers.c
@@ -81,7 +81,7 @@ uint32_t append_oqpsk_chip_seq(uint32_t out_buffer[], uint32_t buffer_start_idx,
  * @param out_buffer_idx the start index to save data in the buffer.
  * @return the next open index in the out_buffer, ie last used index +1, so you can repeatedly call this on a buffer if needed;
  */
-uint32_t data_to_pio_input(const uint8_t in_buffer[], uint32_t in_length, uint32_t out_buffer[], uint32_t out_buffer_idx) {
+uint32_t data_to_pio_input(const uint32_t in_buffer[], uint32_t in_length, uint32_t out_buffer[], uint32_t out_buffer_idx) {
     for (uint32_t i = 0; i < in_length; ++i) {
         
         uint8_t i_byte = in_buffer[i];
